Make bootstrap.bash run only on first boot

+++ bootstrap/bootstrap.bash
@@ -2,5 +2,10 @@
 # Bootstraps a node to work with this repository
 # Assumes it is running on a Debian Jessie distribution
 
-apt-get update
-apt-get install --yes puppet
+
+if [ ! -f /usr/bin/puppet ]; then
+    # This is the first run
+    # Update apt and install puppet!
+    apt-get update
+    apt-get install --yes puppet
+fi
